Rename ObservableArray to TypeScript
