+++ src/observable/index.js
@@ -20,7 +20,7 @@ import {MappedMap} from "./map/MappedMap.js";
 import {JoinedMap} from "./map/JoinedMap.js";
 import {BaseObservableMap} from "./map/BaseObservableMap.js";
 // re-export "root" (of chain) collections
-export { ObservableArray } from "./list/ObservableArray.js";
+export { ObservableArray } from "./list/ObservableArray";
 export { SortedArray } from "./list/SortedArray.js";
 export { MappedList } from "./list/MappedList.js";
 export { AsyncMappedList } from "./list/AsyncMappedList.js";
